muevo contenido Login a clase nueva LoginPanel
